requirements.txt
absl-py==0.7.1
aniso8601==6.0.0
args==0.1.0
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.7.1
bs4==0.0.1
CacheControl==0.12.5
cachy==0.2.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cleo==0.6.8
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.0
docopt==0.6.2
ffmpeg==1.4
Flask==1.0.2
Flask-RESTful==0.3.7
gast==0.2.2
greenlet==0.4.15
grpcio==1.19.0
h5py==2.9.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.2.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lockfile==0.12.2
lxml==4.3.4
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mock==2.0.0
moviepy==1.0.0
msgpack==0.6.1
neovim==0.3.1
numpy==1.16.2
opencv-python==4.0.0.21
parsel==1.5.1
pastel==0.1.0
pbr==5.1.3
Pillow==6.0.0
pipreqs==0.4.9
pkginfo==1.5.0.1
poetry==0.12.11
proglog==0.1.9
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylev==1.3.0
pynvim==0.3.2
pyOpenSSL==19.0.0
pyparsing==2.3.1
PyPrind==2.11.2
pyrsistent==0.14.11
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.1
PyYAML==5.1
queuelib==1.5.0
requests==2.21.0
requests-toolbelt==0.8.0
scikit-learn==0.20.3
scipy==1.2.1
Scrapy==1.6.0
selenium==3.141.0
service-identity==18.1.0
shellingham==1.3.0
six==1.12.0
sklearn==0.0
soupsieve==1.9
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tomlkit==0.5.3
tqdm==4.31.1
Twisted==19.2.1
urllib3==1.24.1
w3lib==1.20.0
webencodings==0.5.1
Werkzeug==0.15.2
yarg==0.1.9
zope.interface==4.6.0
